+++ src/main/java/com/thezz9/schedulerjpa/api/user/service/UserServiceImpl.java
@@ -1,10 +1,12 @@
 package com.thezz9.schedulerjpa.api.user.service;
 
 import com.thezz9.schedulerjpa.api.user.dto.UserCreateRequestDto;
+import com.thezz9.schedulerjpa.api.user.dto.UserDeleteRequestDto;
 import com.thezz9.schedulerjpa.api.user.dto.UserResponseDto;
 import com.thezz9.schedulerjpa.api.user.dto.UserUpdateRequestDto;
 import com.thezz9.schedulerjpa.api.user.entity.User;
 import com.thezz9.schedulerjpa.api.user.repository.UserRepository;
+import com.thezz9.schedulerjpa.common.config.PasswordEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -18,10 +20,12 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
 
     @Override
     public UserResponseDto createUser(UserCreateRequestDto dto) {
-        User savedUser = userRepository.save(new User(dto.getUsername(), dto.getEmail(), dto.getPassword()));
+        String encodedPassword = passwordEncoder.encode(dto.getPassword());
+        User savedUser = userRepository.save(new User(dto.getUsername(), dto.getEmail(), encodedPassword));
         return new UserResponseDto(savedUser);
     }
 
@@ -42,14 +46,24 @@ public class UserServiceImpl implements UserService {
     public UserResponseDto updateUser(Long id, UserUpdateRequestDto dto) {
         User user = userRepository.findById(id).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.NOT_FOUND, "아이디 " + id + "에 해당하는 사용자가 존재하지 않습니다."));
+
+        if (passwordEncoder.matches(dto.getPassword(), user.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
+        }
+
         user.updateUser(dto);
         return new UserResponseDto(user);
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, UserDeleteRequestDto dto) {
         User user = userRepository.findById(id).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.NOT_FOUND, "아이디 " + id + "에 해당하는 사용자가 존재하지 않습니다."));
+
+        if (passwordEncoder.matches(dto.getPassword(), user.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
+        }
+
         userRepository.delete(user);
     }
